refactor(bpmn): extract lint internals to lint.ts; split collect/render
c8ctl-plugin.ts shrinks from ~480 lines to ~140 and is now strictly
host glue: imports, injectFlagsIntoArgs, the bpmnHandler dispatch (with
inlined subcommand-hint logic), and the metadata + commands API
surface. Everything bpmnlint-driven moved to a new lint.ts:

- readBpmnInput (file / stdin async iteration)
- detectCamundaCloudVersion + resolveCamundaCompatConfig + buildLintConfig
- formatElementRef and the isRecord helper
- the lintSubcommand entry point itself

Also addressed the rendering split. The old formatLintResults always
built the text `lines[]` even in JSON mode. Now:

- collectLintResult(results) is the pure data step — returns `{ rows,
  issues, errorCount, warningCount }`. Called once.
- renderLintJson(logger, file, result) emits the JSON shape.
- renderLintText(logger, source, result, { quiet }) builds the
  column-aligned table and summary line; only runs in text mode.

lintSubcommand dispatches to the right renderer based on
c8ctl.outputMode, so terminal-formatting work is skipped entirely when
the user passes --json.
